APPPOCTOOL-10: remove redundant logic
