+++ src/test/java/org/folio/entitlement/it/EntitlementApplicationServiceIT.java
@@ -4,7 +4,7 @@ import static java.net.URLEncoder.encode;
 import static java.net.http.HttpRequest.BodyPublishers.ofString;
 import static java.nio.charset.StandardCharsets.UTF_8;
 import static java.util.Objects.requireNonNull;
-import static org.folio.entitlement.support.TestConstants.DUMMY_SSL_CONTEXT;
+import static org.folio.entitlement.support.TestConstants.HTTP_CLIENT_DUMMY_SSL;
 import static org.folio.entitlement.support.TestUtils.OBJECT_MAPPER;
 import static org.folio.test.security.TestJwtGenerator.generateJwtToken;
 import static org.hamcrest.Matchers.is;
@@ -18,7 +18,6 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 import java.net.URI;
-import java.net.http.HttpClient;
 import java.net.http.HttpRequest;
 import java.net.http.HttpResponse.BodyHandlers;
 import java.util.AbstractMap.SimpleImmutableEntry;
@@ -35,7 +34,6 @@ import org.folio.test.extensions.EnableKeycloakTlsMode;
 import org.folio.test.extensions.KeycloakRealms;
 import org.folio.test.extensions.WireMockStub;
 import org.folio.test.types.IntegrationTest;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.test.context.jdbc.Sql;
@@ -57,15 +55,7 @@ class EntitlementApplicationServiceIT extends BaseIntegrationTest {
     TEST_TENANT, new SimpleImmutableEntry<>("test-login-application", "test-login-application-secret"),
     TEST_TENANT2, new SimpleImmutableEntry<>("test2-login-application", "test2-login-application-secret"));
 
-  private static HttpClient httpClient;
-
   @Autowired private KeycloakProperties keycloakProperties;
-
-  @BeforeEach
-  @SneakyThrows
-  void setUp() {
-    httpClient = HttpClient.newBuilder().sslContext(DUMMY_SSL_CONTEXT).build();
-  }
 
   @Test
   @WireMockStub(scripts = {
@@ -150,7 +140,7 @@ class EntitlementApplicationServiceIT extends BaseIntegrationTest {
       .header("Content-Type", APPLICATION_FORM_URLENCODED_VALUE)
       .build();
 
-    var response = httpClient.send(request, BodyHandlers.ofString(UTF_8));
+    var response = HTTP_CLIENT_DUMMY_SSL.send(request, BodyHandlers.ofString(UTF_8));
     var keycloakTokenJson = OBJECT_MAPPER.readTree(response.body());
     return keycloakTokenJson.path("access_token").asText();
   }
@@ -162,4 +152,3 @@ class EntitlementApplicationServiceIT extends BaseIntegrationTest {
       .collect(Collectors.joining("&"));
   }
 }
-
