+++ src/test/java/org/folio/entitlement/support/TestConstants.java
@@ -1,20 +1,13 @@
 package org.folio.entitlement.support;
 
-import static javax.net.ssl.SSLContext.getInstance;
+import static org.folio.entitlement.support.TestUtils.httpClientWithDummySslContext;
 import static org.folio.integration.kafka.KafkaUtils.getEnvTopicName;
 import static org.folio.integration.kafka.KafkaUtils.getTenantTopicName;
 
-import java.net.Socket;
-import java.security.SecureRandom;
-import java.security.cert.X509Certificate;
+import java.net.http.HttpClient;
 import java.util.UUID;
-import javax.net.ssl.SSLContext;
-import javax.net.ssl.SSLEngine;
-import javax.net.ssl.TrustManager;
-import javax.net.ssl.X509ExtendedTrustManager;
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
-import lombok.SneakyThrows;
 
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
 public class TestConstants {
@@ -36,7 +29,7 @@ public class TestConstants {
   public static final UUID APPLICATION_FLOW_ID = UUID.randomUUID();
   public static final String FLOW_STAGE_ID = FLOW_ID + "/appi-l0/" + APPLICATION_FLOW_ID;
 
-  public static final SSLContext DUMMY_SSL_CONTEXT = dummySslContext();
+  public static final HttpClient HTTP_CLIENT_DUMMY_SSL = httpClientWithDummySslContext();
 
   public static String entitlementTopic() {
     return getEnvTopicName("entitlement");
@@ -52,44 +45,5 @@ public class TestConstants {
 
   public static String systemUserTenantTopic() {
     return getTenantTopicName("mgr-tenant-entitlements.system-user", TENANT_NAME);
-  }
-
-  @SneakyThrows
-  private static SSLContext dummySslContext() {
-    var dummyTrustManager = new X509ExtendedTrustManager() {
-
-      @Override
-      public void checkClientTrusted(X509Certificate[] chain, String authType, Socket socket) {
-      }
-
-      @Override
-      public void checkClientTrusted(X509Certificate[] chain, String authType, SSLEngine engine) {
-      }
-
-      @Override
-      public void checkClientTrusted(X509Certificate[] chain, String authType) {
-      }
-
-      @Override
-      public void checkServerTrusted(X509Certificate[] chain, String authType, Socket socket) {
-      }
-
-      @Override
-      public void checkServerTrusted(X509Certificate[] chain, String authType, SSLEngine engine) {
-      }
-
-      @Override
-      public void checkServerTrusted(X509Certificate[] chain, String authType) {
-      }
-
-      @Override
-      public X509Certificate[] getAcceptedIssuers() {
-        return new X509Certificate[0];
-      }
-    };
-
-    var sslContext = getInstance("TLS");
-    sslContext.init(null, new TrustManager[] {dummyTrustManager}, new SecureRandom());
-    return sslContext;
   }
 }
